Refactored LinksSetEditor into separate components
